feat: change band to take orientation
BREAKING CHANGE: If orient is horizontal, from/toValues are scaled using yScale and the range of the xScale is used as the width of the band. If orient is horizontal, from/toValues are scaled using xScale and the range of the yScale is used as the height of the band.