feat: improved series bar width concepts
fixes #44 #26

BREAKING CHANGE: fractionalBarWidth has been replaced with autoBandwidth

The previous fractionalBarWidth concept was problematic in that it required the series to supply all of the data to the barWidth property (function) in order to support its behaviour. This was not good from a performance perspective, and inflexible.

Following this change, series now expose a bandwidth property (which reflects the same property exposed by d3 band / point scales). This property specifies the width in screen values (pixels).

For example, to create bars with a width of 50 pixels:

```
var svgBar = fc.seriesSvgBar()
    .crossValue(function(d) { return d.x; })
    .mainValue(function(d) { return d.y; })
    .width(50);
```

You can also provide a function that allows configuration of the bandwidth on a per-datapoint basis:

```
var svgBar = fc.seriesSvgBar()
    .crossValue(function(d) { return d.x; })
    .mainValue(function(d) { return d.y; })
    .width(function(d, i) { return i % 2 * 10; });
```

In order to automatically compute the width based on the space between datapoints, you can adapt a series with the autoBandwith component:

```
var svgBar = fc.autoBandwidth(fc.seriesSvgBar())
    .crossValue(function(d) { return d.x; })
    .mainValue(function(d) { return d.y; })
    .widthFraction(0.5);
```

This is equivalent to the previous code:

```
var svgBar = fc.seriesSvgBar()
    .crossValue(function(d) { return d.x; })
    .mainValue(function(d) { return d.y; })
    .barWidth(fc.fractionalBarWidth(0.5));
```

Finally, the autoBandwidth, when used in conjunction with scales that expose a bandwidth property (band / point), this value is propagated to the series.